+++ ormar/__init__.py
@@ -1,6 +1,18 @@
-from ormar.decorators import property_field
-from ormar.exceptions import ModelDefinitionError, MultipleMatches, NoMatch
+from ormar.decorators import (
+    post_delete,
+    post_save,
+    post_update,
+    pre_delete,
+    pre_save,
+    pre_update,
+    property_field,
+)
 from ormar.protocols import QuerySetProtocol, RelationProtocol  # noqa: I100
+from ormar.exceptions import (  # noqa: I100
+    ModelDefinitionError,
+    MultipleMatches,
+    NoMatch,
+)
 from ormar.fields import (  # noqa: I100
     BigInteger,
     Boolean,
@@ -22,6 +34,7 @@ from ormar.models import Model
 from ormar.models.metaclass import ModelMeta
 from ormar.queryset import QuerySet
 from ormar.relations import RelationType
+from ormar.signals import Signal
 
 
 class UndefinedType:  # pragma no cover
@@ -31,7 +44,7 @@ class UndefinedType:  # pragma no cover
 
 Undefined = UndefinedType()
 
-__version__ = "0.6.2"
+__version__ = "0.7.0"
 __all__ = [
     "Integer",
     "BigInteger",
@@ -59,4 +72,11 @@ __all__ = [
     "RelationProtocol",
     "ModelMeta",
     "property_field",
+    "post_delete",
+    "post_save",
+    "post_update",
+    "pre_delete",
+    "pre_save",
+    "pre_update",
+    "Signal",
 ]