+++ ormar/decorators/__init__.py
@@ -1,5 +1,19 @@
 from ormar.decorators.property_field import property_field
+from ormar.decorators.signals import (
+    post_delete,
+    post_save,
+    post_update,
+    pre_delete,
+    pre_save,
+    pre_update,
+)
 
 __all__ = [
     "property_field",
+    "post_delete",
+    "post_save",
+    "post_update",
+    "pre_delete",
+    "pre_save",
+    "pre_update",
 ]